Move `StandStorage` under the `stand` package
